+++ tests/inference/dump-as-vector.R
@@ -1,0 +1,1 @@
+v <- as.vector(1L, mode = "list")